💄 Rename client and improve error handling for feed.py
