+++ talkytrend/handler/alphavantage.py
@@ -1,7 +1,7 @@
 from alpha_vantage.alphaintelligence import AlphaIntelligence
 from loguru import logger
 
-from .client import Client
+from ._client import Client
 
 
 class AlphavantageHandler(Client):
