+++ talkytrend/handler/finnhub.py
@@ -1,6 +1,6 @@
 import finnhub
 
-from .client import Client
+from ._client import Client
 
 
 class FinnhubHandler(Client):
