+++ talkytrend/handler/forexnewsapi.py
@@ -1,7 +1,7 @@
 import aiohttp
 from loguru import logger
 
-from .client import Client
+from ._client import Client
 
 
 class ForexnewsapiHandler(Client):
