+++ talkytrend/handler/newsfilterio.py
@@ -1,6 +1,6 @@
 # from loguru import logger
 
-# from .client import Client
+# from ._client import Client
 
 
 # class NewsfilterioHandler(Client):
